requirements.txt
django==1.2.5
#psycopg2==2.4
django-registration
django-taggit==0.9.3
#django-threadedcomments==0.5.3
#PIL
#django-piston
django-social-auth
python-openid
oauth2
django-smtp-ssl
